requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
curl_cffi==0.6.2
dataclasses-json==0.6.4
dnspython==2.6.1
duckduckgo_search==5.3.0
faiss-cpu==1.8.0
filelock==3.13.4
Flask==3.0.2
frozenlist==1.4.1
fsspec==2024.3.1
google-ai-generativelanguage==0.6.1
google-api-core==2.17.1
google-api-python-client==2.125.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-generativeai==0.5.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.0
grpcio-status==1.62.0
httplib2==0.22.0
huggingface-hub==0.22.2
idna==3.6
importlib-metadata==7.0.1
inflect==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-google-genai==1.0.2
langchain-mongodb==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.47
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.0
packaging==23.2
pandas==2.2.1
Pillow==10.0.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.6.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
setuptools==68.2.2
six==1.16.0
SQLAlchemy==2.0.29
tenacity==8.2.3
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.1
wheel==0.41.2
yarl==1.9.4
zipp==3.17.0
